pyproject.toml
[project]
name = "vibe-learning-rag"
version = "0.1.0"
description = "RAG Development Learning Repository with atomized knowledge points"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "sentence-transformers>=2.2.0",
    "chromadb>=0.4.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "pypdf>=3.0.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.12.0",
    "numpy>=2.4.2",
    "transformers>=5.1.0",
    "torch>=2.10.0",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "anthropic>=0.79.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipython>=8.12.0",
]
